Support network webpage
The support network webpage has been finished with the login system included. The login system now has the feature to log the user out of the website. The user is now able to add and remove people from their support network. Additional table columns added to the SUPPORT_EMAILS table in the database are added. There is also a temporary navigation bar.