First prototype
The user is able to create an account and login. From there they can access the forum (which has some temporary placeholder subforums until we can add the ability to create a subforum). The user can then make a post by clicking add new post in either the subforums and view them (no ability to like/comment yet). They can also use the support network to add/remove people to/from their support network although it is unable to send emails on user progression on the site yet. The leaderboard can be used but you have to test it by manually creating values in the database. It should show the top 10 highest scorers and ignore anyone with 0 points.